% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntm.R
\name{ntm}
\alias{ntm}
\title{ntm}
\usage{
ntm(x)
}
\arguments{
\item{x}{a dataframe containing NAMM, PONS, Lrr and Lat columns}
}
\value{
a tibble with probabilities for each tranquility score, a tranquility score and a tranquility description
}
\description{
Calculate tranquility scores for sites in a dataframe
}
\details{
Your dataframe must contain columns with the attributes NAMM, PONS, Lrr and Lat
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE
 df <- data(ntm)
 ntm(df)
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{case_when}}
}
