% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntm_score.R
\name{ntm_score}
\alias{ntm_score}
\title{FUNCTION_TITLE}
\usage{
ntm_score(NAMM, PONS, Lrr, Lat)
}
\arguments{
\item{NAMM}{PARAM_DESCRIPTION}

\item{PONS}{PARAM_DESCRIPTION}

\item{Lrr}{PARAM_DESCRIPTION}

\item{Lat}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{as_tibble}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{case_when}}
}
