% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntm.R
\name{ntm}
\alias{ntm}
\title{FUNCTION_TITLE}
\usage{
ntm(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{case_when}}
}
