% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/ntm.R
\docType{data}
\name{ntm}
\alias{ntm}
\title{NAMM, PONS, Lrr and Lat values for 5 sites}
\format{
A data frame with 10 rows and 6 variables:
\describe{
  \item{NAMM}{NAMM score, from 1 -5}
  \item{PONS}{PONS score, percentage natural sound}
  \item{Lrr}{Lrr, road and rail noise in decibels}
  \item{Lat}{Lat, corrected overal measure in decibels}
  ...
}
}
\usage{
ntm

ntm(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
A dataset containing NAMM, PONS, Lrr and Lat values for five sites, with and
without aircraft noise included.  These data are taken from page 95 of 
Bentley (2019) Tranquil Spaces. Sharps Redmore Press.

FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{case_when}}
}
\keyword{datasets}
