% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntm_score.R
\name{ntm_score}
\alias{ntm_score}
\title{ntm_score}
\usage{
ntm_score(NAMM, PONS, Lrr, Lat)
}
\arguments{
\item{NAMM}{a NAMM score between 1-5}

\item{PONS}{a PONS score, the percentage of natural sound}

\item{Lrr}{a road and rail score in decibels, 15-90 dB}

\item{Lat}{a corrected overall level, in decibels, 15-90dB}
}
\value{
a tranquility description as a string
}
\description{
calculates a tranquility score from four inputs
}
\details{
each argument must be an integer
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE
 ntm_score(5, 97, 15, 48)
 }
}
}
\seealso{
\code{\link[tibble]{as_tibble}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{case_when}}
}
